Also move turning angle tests to test_path.py
